# acp-writer/api/openapi.yaml
openapi: 3.1.0
info:
  title: ACP Writer API
  description: |
    Composes patient-specific, FHIR-compliant care plans by combining clinical
    decision logic (DMN) with patient data. Manages an internal decision engine
    (Drools/Kogito) and knowledge base (vector store).

    Callers provide patient data directly — the acp-writer does not query FHIR
    servers. The decision engine and vector store are internal implementation
    details hidden behind this API.

    Contract version: 1.0
  version: 0.2.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8082
    description: Local development
paths:
...
  /api/v1/guidelines:
    post:
      operationId: registerGuideline
      summary: Register a clinical practice guideline
      description: |
        Registers guideline-level metadata. Both decision models and
        recommendations reference this by cpg_id.
      tags: [guidelines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CPGMetadata"
      responses:
        "201":
          description: Guideline registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPGMetadata"
        "400":
          description: Invalid metadata

    get:
      operationId: listGuidelines
      summary: List registered guidelines
      tags: [guidelines]
      responses:
        "200":
          description: List of guidelines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CPGMetadata"
  /api/v1/guidelines/{cpgId}:
    parameters:
      - name: cpgId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getGuideline
      summary: Retrieve a registered guideline
      tags: [guidelines]
      responses:
        "200":
          description: Guideline details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CPGMetadata"
        "404":
          description: Guideline not found

    delete:
      operationId: removeGuideline
      summary: Remove a guideline and all associated artifacts
      description: |
        Cascade-deletes all decision models and recommendations associated
        with this guideline.
      tags: [guidelines]
      responses:
        "204":
          description: Guideline and associated artifacts removed
        "404":
          description: Guideline not found

  # --- Decision Models ---
  /api/v1/decisions/models:
    post:
      operationId: deployDecisionModel
      summary: Deploy a DMN decision model
      description: |
        Accepts a DMN 1.4 XML document and deploys it to the internal decision
        engine. The model becomes available for care plan generation and
        direct evaluation.
      tags: [decisions]
      parameters:
        - name: source_cpg
          in: query
          required: false
          schema:
            type: string
        - name: validate_only
          in: query
          required: false
          description: Validate the DMN with the engine without storing it.
          schema:
            type: boolean
            default: false
        - name: replace
          in: query
          required: false
          description: Replace an existing model when its source CPG differs.
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
              description: DMN 1.4 XML document
      responses:
        "201":
          description: Model deployed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionModelSummary"
        "400":
          description: Invalid DMN XML
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: DMN failed engine validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DMNValidationError"
        "409":
          description: Model ID belongs to a different source CPG
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Engine validation unavailable for validate-only requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      operationId: listDecisionModels
      summary: List deployed decision models
      tags: [decisions]
      responses:
        "200":
          description: List of deployed models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DecisionModelSummary"
  /api/v1/decisions/models/{modelId}:
    parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getDecisionModel
      summary: Get a deployed decision model
      tags: [decisions]
      responses:
        "200":
          description: Model details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionModelDetail"
        "404":
          description: Model not found

    delete:
      operationId: removeDecisionModel
      summary: Remove a deployed decision model
      tags: [decisions]
      responses:
        "204":
          description: Model removed
        "404":
          description: Model not found
  /api/v1/decisions/evaluate/{modelId}:
    post:
      operationId: evaluateDecision
      summary: Evaluate a decision model with test inputs
      description: |
        Invokes a deployed DMN model with the provided inputs and returns the
        decision output. Useful for validating extracted DMN before committing.
      tags: [decisions]
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: DMN input values keyed by input data name
              example:
                Systolic BP: 142
                Has Diabetes: true
                Has Kidney Disease: false
      responses:
        "200":
          description: Decision output
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Model not found
        "400":
          description: Invalid input data
        "422":
          description: DMN engine rejected the evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DMNValidationError"

  # --- Knowledge Base (Recommendations) ---
  /api/v1/knowledge/recommendations:
    post:
      operationId: ingestRecommendation
      summary: Ingest a single recommendation
      tags: [knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Recommendation"
      responses:
        "201":
          description: Recommendation ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationSummary"
        "400":
          description: Invalid recommendation

    get:
      operationId: listRecommendations
      summary: List ingested recommendations
      tags: [knowledge]
      parameters:
        - name: source_cpg
          in: query
          schema:
            type: string
          description: Filter by source CPG identifier
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendationSummary"
  /api/v1/knowledge/recommendations/batch:
    post:
      operationId: ingestRecommendationBatch
      summary: Ingest a batch of recommendations
      tags: [knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationBundle"
      responses:
        "201":
          description: Recommendations ingested
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested:
                    type: integer
                  source_cpg:
                    type: string
        "400":
          description: Invalid bundle
  /api/v1/knowledge/recommendations/{recommendationId}:
    parameters:
      - name: recommendationId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getRecommendation
      summary: Retrieve a recommendation by ID
      tags: [knowledge]
      responses:
        "200":
          description: Full recommendation with all metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Recommendation not found
  /api/v1/knowledge/search:
    post:
      operationId: searchRecommendations
      summary: Semantic search over recommendations
      tags: [knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationSearchRequest"
      responses:
        "200":
          description: Search results (summaries, not full objects)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationSearchResponse"

  # --- Care Plans ---
  /api/v1/careplans:
    post:
      operationId: generateCarePlan
      summary: Generate a care plan from patient data
      description: |
        Accepts patient data as a FHIR Bundle or IPS document. Internally
        extracts relevant clinical data, invokes applicable DMN decision
        models, retrieves relevant recommendations from the knowledge base,
        and assembles a FHIR CarePlan Bundle.

        The caller is responsible for providing patient data — this service
        does not query FHIR servers directly.
      tags: [careplans]
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              description: FHIR Bundle or IPS document containing patient data
      responses:
        "201":
          description: Care plan generated
          content:
            application/fhir+json:
              schema:
                type: object
                description: FHIR Bundle containing CarePlan, Goal, and activity resources
        "400":
          description: Invalid patient data
        "422":
          description: Unable to generate care plan (e.g., no applicable decision models)

    get:
      operationId: listCarePlans
      summary: List generated care plans
      tags: [careplans]
      parameters:
        - name: patient
          in: query
          schema:
            type: string
          description: Filter by patient reference
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, approved, rejected]
          description: Filter by care plan status
      responses:
        "200":
          description: List of care plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarePlanSummary"
  /api/v1/careplans/{careplanId}:
    parameters:
      - name: careplanId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getCarePlan
      summary: Retrieve a generated care plan
      tags: [careplans]
      responses:
        "200":
          description: Care plan bundle
          content:
            application/fhir+json:
              schema:
                type: object
                description: FHIR Bundle containing CarePlan and related resources
        "404":
          description: Care plan not found
  /api/v1/careplans/{careplanId}/status:
    put:
      operationId: updateCarePlanStatus
      summary: Update care plan status
      tags: [careplans]
      parameters:
        - name: careplanId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarePlanStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarePlanSummary"
        "404":
          description: Care plan not found

  # --- Health ---
...
components:
  schemas:
...
    CPGMetadata:
      type: object
      required: [cpg_id, title]
      properties:
        contract_version:
          type: string
          default: "1.0"
        cpg_id:
          type: string
        title:
          type: string
        version:
          type: string
        publication_date:
          type: string
          format: date
        evidence_review_date:
          type: string
          format: date
        issuing_body:
          type: string
        grading_system:
          type: string
          enum: [GRADE, COR-LOE, GRADE-COR-hybrid, simplified, verb-implied, ungraded]
        scope:
          type: string
        supersedes:
          type: array
          items:
            type: string

    # --- Decision Models ---
    DecisionModelSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        inputs:
          type: array
          items:
            $ref: "#/components/schemas/DecisionVariable"
        outputs:
          type: array
          items:
            $ref: "#/components/schemas/DecisionVariable"
        deployed_at:
          type: string
          format: date-time
        source_cpg:
          type: string
        category:
          type: string
          enum: [treatment, screening, monitoring, risk-assessment, diagnostic]
        modifies:
          type: array
          items:
            type: string
        source_location:
          $ref: "#/components/schemas/SourceLocation"
    DecisionModelDetail:
      allOf:
        - $ref: "#/components/schemas/DecisionModelSummary"
        - type: object
          properties:
            dmn_xml:
              type: string
...
    DMNValidationError:
      type: object
      required: [error, messages]
      properties:
        error:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DMNValidationMessage"

    # --- Knowledge (Recommendations) ---
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string